+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Implement Kazakus Potion text formatting (@azeier)
 - Implement Jade Golem text formatting (@azeier)
 
+### Changed
+- Change `Launcher.fromUrl` to return launcher instance
+
 ## [0.9.1] - 2016-11-19
 ### Fixed
 - Fix intermediate game states getting lost